requirements.txt
pygame
pytest-approvaltests
